architecture(dream-cycle): extract operating-budget circuit breaker from agentdb-adapter
agentdb-adapter.ts had grown into a 1417-line God-module serving 8 unrelated
bounded contexts (referenced by 8 design docs). Extracts the operating-budget
circuit breaker (~140 lines) into store/operating-budget.ts, re-exported for
zero call-site changes — the same technique bridge-client.ts's own prior
extraction from this file already established. 320/320 tests unchanged
before and after; paperclipai/paperclip (ruClip's architectural model)
already splits its own persistence layer per-domain the same way, including
a separate budget_policies.ts/cost_events.ts.

Dream Cycle 2026-09-02, architecture rotation. Issue #8.
