+++ docs/dream-cycle/LEDGER.md
@@ -1,0 +1,8 @@
+# Dream Cycle Ledger
+
+Durable cross-night memory for the `ruvnet/ruClip` nightly Dream Machine cycle.
+One row per run, appended only, never rewritten.
+
+| Date | Deep | Finding | Issue | PR | Evaluated? | Verdict | Effect | Witness | Prior-night fates |
+|---|---|---|---|---|---|---|---|---|---|
+| 2026-09-02 | architecture | `agentdb-adapter.ts` (1417 lines, 8 bounded contexts) is a God-module; extracted the operating-budget circuit breaker into `store/operating-budget.ts` (re-exported, zero call-site changes) | [#8](https://github.com/ruvnet/ruClip/issues/8) | (draft, see issue) | yes | ACCEPT | -116 lines in the god-module, 320/320 tests unchanged, 0 other files touched | `24e92f50954ff3fc5aa69ec3fc617cfd5cb55a58980e35da86a79fd9d2f04ffb` | n/a (first run) |
